+++ core/docs/changelog.md
@@ -1,5 +1,12 @@
 # TBD
 
+# 1.15.7
+### Changes
+* Upgraded to Kurtosis Client v0.3.0
+
+### Features
+* Add a new method `GetServiceInfo` in API container which can be used to get relevant information about a service running in the network
+
 # 1.15.6
 ### Changes
 * Upgraded to using Kurtosis Client v0.2.2, with the `ExecuteBulkCommands` endpoint